requirements.txt
black==22.12.0
Django==4.1.5
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9.5